# backend/openapi.json
{
  "components": {
    "schemas": {
      "Decision": {
        "properties": {
          "agent_response": {
            "title": "Agent Response",
            "type": "string"
          },
          "anomaly_types": {
            "items": {},
            "title": "Anomaly Types",
            "type": "array"
          },
          "attempt": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Attempt"
          },
          "fixed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fixed"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "number"
          },
          "tools_used": {
            "items": {
              "type": "string"
            },
            "title": "Tools Used",
            "type": "array"
          }
        },
        "required": [
          "id",
          "run_id",
          "timestamp",
          "anomaly_types",
          "tools_used",
          "agent_response"
        ],
        "title": "Decision",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "The single shape every declared error response uses.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "error_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Id"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MetricEntry": {
        "properties": {
          "anomaly_injected": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anomaly Injected"
          },
          "epoch": {
            "title": "Epoch",
            "type": "integer"
          },
          "grad_norm": {
            "title": "Grad Norm",
            "type": "number"
          },
          "step": {
            "title": "Step",
            "type": "integer"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "number"
          },
          "train_loss": {
            "title": "Train Loss",
            "type": "number"
          },
          "val_acc": {
            "title": "Val Acc",
            "type": "number"
          },
          "val_loss": {
            "title": "Val Loss",
            "type": "number"
          }
        },
        "required": [
          "step",
          "epoch",
          "train_loss",
          "val_loss",
          "val_acc",
          "grad_norm",
          "timestamp"
        ],
        "title": "MetricEntry",
        "type": "object"
      },
      "Run": {
        "properties": {
          "config_path": {
            "title": "Config Path",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "metrics_file": {
            "title": "Metrics File",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "training_dir": {
            "title": "Training Dir",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "config_path",
          "metrics_file",
          "training_dir",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "Run",
        "type": "object"
      },
      "RunCreate": {
        "additionalProperties": false,
        "properties": {
          "config_path": {
            "minLength": 1,
            "pattern": "\\S",
            "title": "Config Path",
            "type": "string"
          },
          "metrics_file": {
            "minLength": 1,
            "pattern": "\\S",
            "title": "Metrics File",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "training_dir": {
            "minLength": 1,
            "pattern": "\\S",
            "title": "Training Dir",
            "type": "string"
          }
        },
        "required": [
          "name",
          "config_path",
          "metrics_file",
          "training_dir"
        ],
        "title": "RunCreate",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "AutoDebug API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "root__get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Root"
      }
    },
    "/runs/": {
      "get": {
        "operationId": "get_runs_runs__get",
        "parameters": [
          {
            "description": "maximum runs to return, newest first",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "description": "maximum runs to return, newest first",
              "maximum": 5000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "runs to skip, newest first",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "runs to skip, newest first",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Run"
                  },
                  "title": "Response Get Runs Runs  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the request asked for more rows than one request may read"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the request failed inside the service"
          }
        },
        "summary": "Get Runs",
        "tags": [
          "runs"
        ]
      },
      "post": {
        "operationId": "create_run_runs__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the request body could not be decoded"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the request failed inside the service"
          }
        },
        "summary": "Create Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/runs/{run_id}": {
      "get": {
        "operationId": "get_run_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the addressed run does not exist"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the request failed inside the service"
          }
        },
        "summary": "Get Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/runs/{run_id}/decisions": {
      "get": {
        "operationId": "get_decisions_runs__run_id__decisions_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "description": "maximum rows to return; omit for all of them",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "description": "maximum rows to return; omit for all of them",
              "maximum": 5000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "rows to skip, newest first",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "rows to skip, newest first",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Decision"
                  },
                  "title": "Response Get Decisions Runs  Run Id  Decisions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the addressed run does not exist"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the request asked for more rows than one request may read"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the request failed inside the service"
          }
        },
        "summary": "Get Decisions",
        "tags": [
          "decisions"
        ]
      },
      "post": {
        "operationId": "insert_decision_runs__run_id__decisions_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Decision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Decision"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the request body could not be decoded"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the addressed run does not exist"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the path and the body address different runs"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the request failed inside the service"
          }
        },
        "summary": "Insert Decision",
        "tags": [
          "decisions"
        ]
      }
    },
    "/runs/{run_id}/metrics": {
      "get": {
        "operationId": "get_metrics_runs__run_id__metrics_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "description": "maximum rows to return; omit for the whole series",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "description": "maximum rows to return; omit for the whole series",
              "maximum": 5000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "rows to skip, oldest first",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "rows to skip, oldest first",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MetricEntry"
                  },
                  "title": "Response Get Metrics Runs  Run Id  Metrics Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the addressed run does not exist"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the request asked for more rows than one request may read"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the request failed inside the service"
          }
        },
        "summary": "Get Metrics",
        "tags": [
          "metrics"
        ]
      }
    },
    "/runs/{run_id}/metrics/sync": {
      "post": {
        "operationId": "sync_metrics_runs__run_id__metrics_sync_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the addressed run does not exist"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the run exists but its metrics file is not readable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the request failed inside the service"
          }
        },
        "summary": "Sync Metrics",
        "tags": [
          "metrics"
        ]
      }
    },
    "/runs/{run_id}/status": {
      "patch": {
        "operationId": "update_status_runs__run_id__status_patch",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": true,
            "schema": {
              "enum": [
                "running",
                "completed",
                "failed"
              ],
              "title": "Status",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the request was well formed but not permitted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the addressed run does not exist"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "the request failed inside the service"
          }
        },
        "summary": "Update Status",
        "tags": [
          "runs"
        ]
      }
    }
  }
}